data-collector/requirements.txt
# Core Dependencies
playwright==1.48.0
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0

# AI/NLP
anthropic==0.40.0
nltk==3.8.1
spacy==3.7.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
pydantic==2.5.3
python-dateutil==2.8.2
pytz==2024.1

# Async & Rate Limiting
aiohttp==3.9.1
asyncio==3.4.3
ratelimit==2.2.1

# Logging & Monitoring
loguru==0.7.2
tqdm==4.66.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Scheduling
apscheduler==3.10.4  # 자동 스케줄링

# Optional (for future)
# redis==5.0.1         # 캐싱 (추후 추가)
